Create Hangman
This is a very popular game where the user knows the length of the word to guess and he is supposed to guess single letters of that word. The user can't guess more than 7 letters that aren't in the word to guess. If the user uses 8th wrong letter he'll lose and he'll win if he guesses all letters in word to guess.